spec: fix up licensing
Because rust binaries are statically linked the whole binary
has to take into account the combined licenses of all the
components that go into making that binary. Add the cargo
licensing bits to the package and update the combined license
to reflect the output. The healthchecks are just BSD-3-Clause
so reflect that in the sub-package.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>